OSIT-main/OSIT-main/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
numpy==2.4.4
pandas
joblib==1.4.2
pillow==10.4.0
langchain==0.2.14
langchain-google-genai==1.0.10
google-generativeai==0.7.2
google-genai